fix some error handling so findImages gets executed

+++ src/pkg/generate/common.go
@@ -64,9 +64,11 @@ func writeEmbeddedFolder(folder embed.FS, basePathToRemove string, basePathToAdd
 		panic(err)
 	}
 
-	// Attempt to remove the specified base directory if it is empty
-	err = os.Remove(filepath.Join(config.GenerateOutputDir, basePathToRemove))
-	if err != nil {
-		panic(err)
+	// Attempt to remove the specified base directory if one was specified
+	if basePathToAdd != "" {
+		err = os.Remove(filepath.Join(config.GenerateOutputDir, basePathToRemove))
+		if err != nil {
+			panic(err)
+		}
 	}
 }
